+++ src/lib/store.ts
@@ -15,6 +15,7 @@ export const searchValue = writable('')
 export const localEditStatus = writable<LocalEditStatus>({
   available: false,
   mode: 'readonly',
+  capabilities: [],
 })
 export const whenAppReady: Writable<Promise<void>> = writable(
   new Promise(() => {}),